code

+++ Tree/17_Maximum_Path_sum.cpp
@@ -19,10 +19,10 @@
 using namespace std;
 struct node
 {
-    long long data;
+    int data;
     struct node *left;
     struct node *right;
-    node(long long data)
+    node(int data)
     {
         this->data = data;
         this->left = this->right = NULL;
